Allows tagging queues on creation
Adds tags to queues on creation. There are two ways to do this:

1. Specify the tags in the `create_queue` call.
2. Separately call `tag_queue` after creation.

The first option is clearly more efficient as there's only a single call. Unfortunately a not-entirely-documented 'feature' of SQS is that if you try to create an already-existing queue specifying tags and the tags are different to the existing tags (even if the existing queue has no tags) then the call will fail with an error saying the resource already exists with different tags, rather than re-tagging the queue.

The second approach is less efficient as it requires extra calls, but will add new tags and overwrite any existing tags that have changed (although won't remove any existing tags).

Given this choice I've elected to go with the second option partly because I think it means the behaviour of the library will be closer to the expectations of people using it, and partly because this is a new feature meaning anybody with existing queues won't have tags on them so this will allow them to add tags to those queues without having to recreate them.

For testing I've elected to add a tag to the broker that connects to AWS to exercise the new functionality against real resources, as well as add a test to verify that tags are added.